added new unit testcases for pkgagent, coverage is 64% now, fixed a couple pkgagent bugs found using the unit test.
git-svn-id: http://svn.code.sf.net/p/fossology/code/branches/fossology2.0/fossology@4526 a5643f74-2041-0410-a2b9-94ffef304a07

+++ src/pkgagent/agent_tests/testRecordMetadataDEB.c
@@ -87,6 +87,23 @@ void test_RecordMetadataDEB()
   int Result = RecordMetadataDEB(pi);
   printf("RecordMetadataDEB Result is:%d\n", Result);
 
+  /* Check data correction */
+  memset(SQL,'\0',MAXSQL);
+  snprintf(SQL,MAXSQL,"SELECT pkg_pk, pkg_name, pkg_arch, version, maintainer, description FROM pkg_deb INNER JOIN pfile ON pfile_fk = '%ld' AND pfile_fk = pfile_pk;", pi->pFileFk);
+  result =  PQexec(db_conn, SQL);
+  if (fo_checkPQresult(db_conn, result, SQL, __FILE__, __LINE__))
+  {
+    printf("Get pkg information ERROR!\n");
+    exit(-1);
+  }
+  CU_ASSERT_STRING_EQUAL(PQgetvalue(result, 0, 1), "Test Pkg");
+  CU_ASSERT_STRING_EQUAL(PQgetvalue(result, 0, 2), "Test Arch");
+  CU_ASSERT_STRING_EQUAL(PQgetvalue(result, 0, 3), "Test version");
+  CU_ASSERT_STRING_EQUAL(PQgetvalue(result, 0, 4), "Test maintainer");
+  CU_ASSERT_STRING_EQUAL(PQgetvalue(result, 0, 5), "Test description");
+  PQclear(result);
+
+
   /* Clear testing data in database */
   memset(SQL,'\0',MAXSQL);
   snprintf(SQL,MAXSQL,"DELETE FROM pkg_deb_req WHERE pkg_fk IN (SELECT pkg_pk FROM pkg_deb WHERE pfile_fk = '%ld');", pi->pFileFk);
